backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
mangum==0.18.0

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9

# Data validation
pydantic==2.9.2
pydantic-settings==2.5.2

# AWS
boto3==1.35.36

# HTTP client
httpx==0.27.2

# Auth
python-jose[cryptography]==3.3.0

# Excel processing
openpyxl==3.1.5

# Form data (required by FastAPI file uploads)
python-multipart==0.0.12

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
hypothesis==6.112.1
